Extract spoiler, plain, html, me to messages
